thuc hanh

+++ ss1/change_currency/src/main/java/com/change_currency/controller/Controllers.java
@@ -19,9 +19,9 @@ public class Controllers {
     }
 
     @PostMapping("/")
-    public String convert(Model model, @RequestParam(value = "usd")int usd, @RequestParam(value = "vnd")int vnd){
+    public String convert(Model model, @RequestParam(value = "usd")long usd, @RequestParam(value = "vnd")long vnd){
 
-        int result =  iConvertService.usdToVnd(usd,vnd);
+        long result =  iConvertService.usdToVnd(usd,vnd);
 
             model.addAttribute("vnd",vnd);
             model.addAttribute("usd",usd);
