+++ ss1/change_currency/src/main/java/com/change_currency/service/IConvertService.java
@@ -1,5 +1,5 @@
 package com.change_currency.service;
 
 public interface IConvertService {
-    int usdToVnd(int usd, int vnd);
+    long usdToVnd(long usd, long vnd);
 }
